requirements.txt
numpy>=1.19.0
torch>=1.7.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
PyPDF2>=3.0.0
requests>=2.25.1